Fix metal dot_product kernel

+++ src/metal/dot_product.metal
@@ -1,7 +1,7 @@
 kernel void dot_product(
-  device const float* inA,
-  device const float* inB,
-  device float* result,
+  device const uint* inA,
+  device const uint* inB,
+  device uint* result,
   uint index [[thread_position_in_grid]])
 {
   result[index] = inA[index] * inB[index];
